test(vscode): decouple GradleService from vscode module for unit tests
Move the vscode.workspace.getConfiguration lookup out of GradleService
and inject it as an argsProvider callback from extension.ts. The service
no longer imports vscode, so gradleService.test.ts can load it under
plain mocha without a module-resolution failure.
